backend/requirements.txt
Flask==3.0.3 
Flask-CORS==4.0.1 
Flask-SQLAlchemy==3.1.1 
SQLAlchemy==2.0.31 
ollama==0.3.3 
python-dotenv==1.0.1